Refactor: break LibraryBrowser/PackBrowser coupling, extract shared library UI
PackBrowser imported AssetRow and LOAD_REASON_LABELS back out of its sibling
view LibraryBrowser. Two sibling views importing UI from each other is the
smell; neither owned the shared pieces, so the panel was load-bearing for the
modal's rows.

Lift what both use into modules of their own:

- AssetRow.tsx      — the shared auditionable-sound row
- loadReasons.ts    — LOAD_REASON_LABELS, the one place a load failure is named
- LibraryFacets.tsx — FacetBar/FacetGroup and the TYPE/KIND label tables
- PackList.tsx      — the modal's left rail, which owns its own pack filter

Both views now import from the shared modules instead of from each other.

AssetRow's styles (.library-row, .library-audition, .library-insert) live in
LibraryBrowser.css, and CSS arrives through per-component side-effect imports,
so AssetRow.tsx imports that sheet itself rather than relying on whichever view
mounted first. LibraryFacets needs .library-search from the panel sheet and its
chips/layout from the modal sheet, so it imports both, panel first — preserving
the order that lets `.pack-browser-detail .library-search` override the base.

Pure refactor: no behavior, DOM, or visual change.
