+++ src/library/LibraryBrowser.tsx
@@ -2,10 +2,7 @@ import {
 	HiSolidChevronDown,
 	HiSolidChevronRight,
 	HiSolidExclamationTriangle,
-	HiSolidPlay,
-	HiSolidPlus,
 	HiSolidSquares2x2,
-	HiSolidStop,
 } from "solid-icons/hi";
 import {
 	createEffect,
@@ -19,8 +16,10 @@ import {
 } from "solid-js";
 import type { Analytics } from "../analytics/analytics";
 import TapeLoader from "../components/TapeLoader";
+import AssetRow from "./AssetRow";
 import type { PreviewEngine } from "./audition";
 import type { LibraryClient } from "./libraryClient";
+import { LOAD_REASON_LABELS } from "./loadReasons";
 import type { LibraryAsset, LibraryPackSummary } from "./manifest";
 import PackBrowser from "./PackBrowser";
 import type { LibraryTreeGroup, LibraryTreePack } from "./tree";
@@ -49,13 +48,6 @@ export interface LibraryBrowserProps {
 	 */
 	readonly onPackBrowserOpenChange?: (open: boolean) => void;
 }
-
-export const LOAD_REASON_LABELS: Record<string, string> = {
-	network: "Check your connection.",
-	not_found: "This library is unavailable.",
-	corrupt: "The library data could not be read.",
-	timeout: "The library took too long to load.",
-};
 
 /**
  * The library panel: a narrow, always-available column beside the workspace
@@ -339,69 +331,5 @@ function Chevron(props: { expanded: boolean }): JSX.Element {
 				<HiSolidChevronDown size={12} />
 			</Show>
 		</span>
-	);
-}
-
-/** One auditionable sound. Shared by the panel tree and the pack browser. */
-export function AssetRow(props: {
-	asset: LibraryAsset;
-	active: boolean;
-	error: string | null;
-	/** Shown in the pack browser, where a result can come from any pack. */
-	showPack?: boolean;
-	onPlay: () => void;
-	onStop: () => void;
-	onInsert?: () => void;
-}): JSX.Element {
-	return (
-		<li
-			class="library-row"
-			classList={{
-				"library-row-active": props.active,
-				"library-row-error": props.error !== null,
-			}}
-		>
-			<button
-				type="button"
-				class="library-audition"
-				aria-label={
-					props.active
-						? `Stop ${props.asset.name}`
-						: `Audition ${props.asset.name}`
-				}
-				aria-pressed={props.active}
-				onClick={() => (props.active ? props.onStop() : props.onPlay())}
-			>
-				<Show when={props.active} fallback={<HiSolidPlay size={12} />}>
-					<HiSolidStop size={12} />
-				</Show>
-			</button>
-			<div class="library-row-meta">
-				<span class="library-row-name">{props.asset.name}</span>
-				<Show when={props.showPack || props.error}>
-					<span class="library-row-tags">
-						<Show when={props.showPack}>
-							{props.asset.role} · {props.asset.packName}
-						</Show>
-						<Show when={props.error}>
-							<span class="library-row-error-text">
-								{props.showPack ? " · " : ""}
-								{LOAD_REASON_LABELS[props.error ?? ""] ?? "Could not load."}
-							</span>
-						</Show>
-					</span>
-				</Show>
-			</div>
-			<Show when={props.onInsert}>
-				<button
-					type="button"
-					class="library-insert"
-					aria-label={`Insert ${props.asset.name}`}
-					onClick={() => props.onInsert?.()}
-				>
-					<HiSolidPlus size={12} />
-				</button>
-			</Show>
-		</li>
 	);
 }
