+++ src/library/PackBrowser.tsx
@@ -6,7 +6,6 @@ import {
 import {
 	createEffect,
 	createMemo,
-	createSignal,
 	For,
 	type JSX,
 	onMount,
@@ -19,30 +18,16 @@ import {
 import TapeLoader from "../components/TapeLoader";
 import type { ShortcutContext, ShortcutHandlers } from "../shortcuts";
 import { useShortcuts } from "../shortcuts";
-import { AssetRow, LOAD_REASON_LABELS } from "./LibraryBrowser";
-import type {
-	LibraryAsset,
-	LibraryAssetType,
-	LibraryPackSummary,
-} from "./manifest";
-import { EMPTY_PACK_FILTER, filterPacks, type PackKind } from "./search";
+import AssetRow from "./AssetRow";
+import { FacetBar } from "./LibraryFacets";
+import { LOAD_REASON_LABELS } from "./loadReasons";
+import type { LibraryAsset, LibraryPackSummary } from "./manifest";
+import PackList from "./PackList";
 import type { useLibraryBrowser } from "./useLibraryBrowser";
 import "./PackBrowser.css";
 
 /** A modal suppresses every other shortcut context while it is open (PRD KEY-01). */
 const DIALOG_CONTEXTS: readonly ShortcutContext[] = ["dialog"];
-
-const TYPE_LABELS: Record<LibraryAssetType, string> = {
-	"one-shot": "One-shots",
-	loop: "Loops",
-	preset: "Presets",
-};
-
-const KIND_LABELS: Record<PackKind, string> = {
-	factory: "Factory",
-	"third-party": "Third-party",
-	user: "Yours",
-};
 
 export interface PackBrowserProps {
 	readonly browser: ReturnType<typeof useLibraryBrowser>;
@@ -68,13 +53,15 @@ export interface PackBrowserProps {
  * in the KEY-01 registry, dispatched by the shared controller like every other
  * mapping, which is what keeps the combination written down once.
  *
- * Facets live here, not in the panel: filtering by genre, role and character
- * needs the width, and the panel is 250px.
+ * Facets live on this surface, not in the panel: filtering by genre, role and
+ * character needs the width, and the panel is 250px. This component is the
+ * dialog frame and the choice of what the detail pane shows; the rail that
+ * narrows and picks a pack is {@link PackList}, and the facet controls the
+ * detail pane filters with are `LibraryFacets`.
  */
 export default function PackBrowser(props: PackBrowserProps): JSX.Element {
 	const { browser } = props;
 	const analytics = props.analytics ?? defaultAnalytics;
-	const [packFilter, setPackFilter] = createSignal(EMPTY_PACK_FILTER);
 
 	useShortcuts({
 		handlers: (): ShortcutHandlers => ({
@@ -103,10 +90,6 @@ export default function PackBrowser(props: PackBrowserProps): JSX.Element {
 		void browser.selectPack(first.slug);
 	});
 
-	const visiblePacks = createMemo(() =>
-		filterPacks(browser.packs(), packFilter()),
-	);
-
 	const selectedPack = createMemo<LibraryPackSummary | null>(() => {
 		const slug = browser.selectedPackSlug();
 		if (slug === null) return null;
@@ -115,22 +98,6 @@ export default function PackBrowser(props: PackBrowserProps): JSX.Element {
 
 	const isAdded = (pack: LibraryPackSummary) =>
 		props.addedPackIds.includes(pack.id);
-
-	function toggleKind(kind: PackKind): void {
-		setPackFilter((prev) => ({
-			...prev,
-			kinds: prev.kinds.includes(kind)
-				? prev.kinds.filter((existing) => existing !== kind)
-				: [...prev.kinds, kind],
-		}));
-	}
-
-	const kinds = createMemo<PackKind[]>(() => {
-		const present = new Set(browser.packs().map((pack) => pack.kind));
-		return (["factory", "third-party", "user"] as const).filter((kind) =>
-			present.has(kind),
-		);
-	});
 
 	return (
 		<div class="pack-browser">
@@ -168,101 +135,7 @@ export default function PackBrowser(props: PackBrowserProps): JSX.Element {
 				</header>
 
 				<div class="pack-browser-body">
-					<nav class="pack-browser-list" aria-label="Packs">
-						<input
-							type="search"
-							class="pack-browser-search"
-							placeholder="Search packs"
-							aria-label="Search packs"
-							value={packFilter().query}
-							onInput={(event) =>
-								setPackFilter((prev) => ({
-									...prev,
-									query: event.currentTarget.value,
-								}))
-							}
-						/>
-						<div class="pack-browser-kinds">
-							<For each={kinds()}>
-								{(kind) => {
-									const active = createMemo(() =>
-										packFilter().kinds.includes(kind),
-									);
-									return (
-										<button
-											type="button"
-											class="library-chip"
-											classList={{ "library-chip-active": active() }}
-											aria-pressed={active()}
-											onClick={() => toggleKind(kind)}
-										>
-											{KIND_LABELS[kind]}
-										</button>
-									);
-								}}
-							</For>
-						</div>
-						<ul class="pack-browser-packs">
-							<li>
-								<button
-									type="button"
-									class="pack-browser-pack"
-									classList={{
-										"pack-browser-pack-active":
-											browser.selectedPackSlug() === null,
-									}}
-									aria-pressed={browser.selectedPackSlug() === null}
-									onClick={() => void browser.selectPack(null)}
-								>
-									<span class="pack-browser-pack-name">All sounds</span>
-									<span class="pack-browser-pack-meta">Every pack</span>
-								</button>
-							</li>
-							<For each={visiblePacks()}>
-								{(pack) => {
-									const failed = createMemo(() =>
-										browser
-											.packErrors()
-											.some((error) => error.packSlug === pack.slug),
-									);
-									return (
-										<li>
-											<button
-												type="button"
-												class="pack-browser-pack"
-												classList={{
-													"pack-browser-pack-active":
-														browser.selectedPackSlug() === pack.slug,
-													"pack-browser-pack-failed": failed(),
-												}}
-												aria-pressed={browser.selectedPackSlug() === pack.slug}
-												onClick={() => void browser.selectPack(pack.slug)}
-											>
-												<span class="pack-browser-pack-name">{pack.name}</span>
-												<span class="pack-browser-pack-meta">
-													{pack.publisher} ·{" "}
-													{failed()
-														? "unavailable"
-														: `${pack.assetCount} sounds`}
-												</span>
-												<Show when={isAdded(pack)}>
-													<span class="pack-browser-pack-added">
-														<HiSolidCheckCircle size={14} />
-														<span class="visually-hidden">Added</span>
-													</span>
-												</Show>
-											</button>
-										</li>
-									);
-								}}
-							</For>
-							<Show when={visiblePacks().length === 0}>
-								<li class="pack-browser-no-packs">
-									No packs match. Clear the search to see them all.
-								</li>
-							</Show>
-						</ul>
-					</nav>
+					<PackList browser={browser} isAdded={isAdded} />
 
 					<section class="pack-browser-detail" aria-label="Pack">
 						<Show
@@ -430,101 +303,6 @@ function Results(props: {
 					</For>
 				</ul>
 			</Show>
-		</Show>
-	);
-}
-
-/** The active-facet controls, computed from the assets actually present. */
-function FacetBar(props: {
-	browser: ReturnType<typeof useLibraryBrowser>;
-}): JSX.Element {
-	const { browser } = props;
-	const anyFilter = createMemo(() => {
-		const filter = browser.filter();
-		return (
-			filter.genres.length > 0 ||
-			filter.roles.length > 0 ||
-			filter.characters.length > 0 ||
-			filter.types.length > 0 ||
-			filter.query.trim() !== ""
-		);
-	});
-	return (
-		<div class="library-facets">
-			<input
-				type="search"
-				class="library-search"
-				placeholder="Search sounds"
-				aria-label="Search sounds"
-				value={browser.filter().query}
-				onInput={(event) => browser.setQuery(event.currentTarget.value)}
-			/>
-			<FacetGroup
-				label="Type"
-				values={browser.facets().types}
-				active={browser.filter().types}
-				display={(type) => TYPE_LABELS[type as LibraryAssetType]}
-				onToggle={(type) => browser.toggleType(type as LibraryAssetType)}
-			/>
-			<FacetGroup
-				label="Genre"
-				values={browser.facets().genres}
-				active={browser.filter().genres}
-				onToggle={(genre) => browser.toggleGenre(genre)}
-			/>
-			<FacetGroup
-				label="Role"
-				values={browser.facets().roles}
-				active={browser.filter().roles}
-				onToggle={(role) => browser.toggleRole(role)}
-			/>
-			<FacetGroup
-				label="Character"
-				values={browser.facets().characters}
-				active={browser.filter().characters}
-				onToggle={(character) => browser.toggleCharacter(character)}
-			/>
-			<Show when={anyFilter()}>
-				<button
-					type="button"
-					class="library-clear"
-					onClick={() => browser.clearFilters()}
-				>
-					Clear filters
-				</button>
-			</Show>
-		</div>
-	);
-}
-
-function FacetGroup(props: {
-	label: string;
-	values: readonly string[];
-	active: readonly string[];
-	display?: (value: string) => string;
-	onToggle: (value: string) => void;
-}): JSX.Element {
-	return (
-		<Show when={props.values.length > 0}>
-			<fieldset class="library-facet-group">
-				<legend class="library-facet-label">{props.label}</legend>
-				<For each={props.values}>
-					{(value) => {
-						const isActive = createMemo(() => props.active.includes(value));
-						return (
-							<button
-								type="button"
-								class="library-chip"
-								classList={{ "library-chip-active": isActive() }}
-								aria-pressed={isActive()}
-								onClick={() => props.onToggle(value)}
-							>
-								{props.display ? props.display(value) : value}
-							</button>
-						);
-					}}
-				</For>
-			</fieldset>
 		</Show>
 	);
 }
